Moving stuff to more appropriate places
