+++ impl11/ddraw/PrimarySurface.cpp
@@ -6,7 +6,6 @@
 #include <wincodec.h>
 
 #include "common.h"
-#include "globals.h"
 #include "DeviceResources.h"
 #include "PrimarySurface.h"
 #include "BackbufferSurface.h"
@@ -19,33 +18,9 @@
 #include "XwaDrawRadarHook.h"
 #include "XwaDrawBracketHook.h"
 #include "effects.h"
+#include "globals.h"
 #include "commonVR.h"
-#include "SteamVR.h"
-#include "DirectSBS.h"
 #include "VRConfig.h"
-
-#define DBG_MAX_PRESENT_LOGS 0
-
-#include <vector>
-
-extern uint32_t* g_playerIndex;
-extern uint32_t* g_playerInHangar;
-bool g_bPrevPlayerInHangar = false;
-#define GENERIC_POV_SCALE 44.0f
-// Set to true in PrimarySurface Present 2D (Flip)
-extern bool g_bInTechRoom;
-bool g_bKeybExitHyperspace = true;
-
-/*
-dword& s_V0x09C6E38 = *(dword*)0x009C6E38;
-When the value is different of 0xFFFF, the player craft is in a hangar.
-*/
-
-bool g_bMetricParamsNeedReapply = false;
-bool LoadFocalLength();
-void ApplyFocalLength(float focal_length);
-void SaveFocalLength();
-Matrix4 g_ReflRotX;
 
 // Text Rendering
 TimedMessage g_TimedMessages[MAX_TIMED_MESSAGES];
@@ -63,49 +38,6 @@ bool rayTriangleIntersect(
 	const Vector3 &orig, const Vector3 &dir,
 	const Vector3 &v0, const Vector3 &v1, const Vector3 &v2,
 	float &t, Vector3 &P, float &u, float &v);
-
-
-
-void InGameToScreenCoords(UINT left, UINT top, UINT width, UINT height, float x, float y, float *x_out, float *y_out);
-void ScreenCoordsToInGame(float left, float top, float width, float height, float x, float y, float *x_out, float *y_out);
-void GetScreenLimitsInUVCoords(float *x0, float *y0, float *x1, float *y1, bool UseNonVR=false);
-
-
-bool g_bTogglePostPresentHandoff = false, g_bInTechRoom = false, g_bSteamVRMirrorWindowLeftEye = true;
-
-// V0x00782848
-DWORD *XwaGlobalLightsCount = (DWORD *)0x00782848;
-
-// V0x007D4FA0
-XwaGlobalLight *XwaGlobalLights = (XwaGlobalLight *)0x007D4FA0; // Maximum 8 lights
-
-struct XwaVector3
-{
-	/* 0x0000 */ float x;
-	/* 0x0004 */ float y;
-	/* 0x0008 */ float z;
-};
-
-// S0x0000002
-struct XwaMatrix3x3
-{
-	/* 0x0000 */ float _11;
-	/* 0x0004 */ float _12;
-	/* 0x0008 */ float _13;
-	/* 0x000C */ float _21;
-	/* 0x0010 */ float _22;
-	/* 0x0014 */ float _23;
-	/* 0x0018 */ float _31;
-	/* 0x001C */ float _32;
-	/* 0x0020 */ float _33;
-};
-
-// S0x0000012
-struct XwaTransform
-{
-	/* 0x0000 */ XwaVector3 Position;
-	/* 0x000C */ XwaMatrix3x3 Rotation;
-};
 
 // S0x0000001
 // L00439B30
@@ -335,232 +267,6 @@ struct MainVertex
 		tex[0] = tx; tex[1] = ty;
 	}
 };
-
-extern bool g_bBloomEnabled, g_bAOEnabled, g_bApplyXWALightsIntensity, g_bProceduralSuns, g_b3DSunPresent, g_b3DSkydomePresent;
-extern float g_fBloomAmplifyFactor, g_fHangarAmbient, g_fGlobalAmbient;
-extern float g_fSpecIntensity, g_fSpecBloomIntensity, g_fXWALightsSaturation, g_fXWALightsIntensity;
-
-extern float g_fHDRLightsMultiplier, g_fHDRWhitePoint;
-extern bool g_bHDREnabled;
-bool g_bGlobalSpecToggle = true;
-//extern float g_fFlareAspectMult;
-
-extern float g_fConcourseScale, g_fConcourseAspectRatio;
-extern int   g_iDraw2DCounter;
-extern bool  g_bStartedGUI, g_bPrevStartedGUI, g_bIsScaleableGUIElem, g_bPrevIsScaleableGUIElem, g_bScaleableHUDStarted, g_bDynCockpitEnabled;
-extern bool g_bDCWasClearedOnThisFrame;
-extern bool  g_bEnableVR, g_bDisableBarrelEffect;
-extern bool  g_bDumpGUI;
-extern int   g_iDrawCounter, g_iExecBufCounter, g_iPresentCounter, g_iNonZBufferCounter, g_iDrawCounterAfterHUD;
-extern bool  g_bTargetCompDrawn, g_bPrevIsFloatingGUI3DObject, g_bIsFloating3DObject;
-extern unsigned int g_iFloatingGUIDrawnCounter;
-
-bool g_bRendering3D = false; // Set to true when the system is about to render in 3D
-
-// SteamVR
-extern DWORD g_FullScreenWidth, g_FullScreenHeight;
-extern Matrix4 g_viewMatrix;
-extern VertexShaderMatrixCB g_VSMatrixCB;
-
-extern bool g_bStart3DCapture, g_bDo3DCapture;
-extern FILE *g_HackFile;
-#ifdef DBR_VR
-extern bool g_bCapture2DOffscreenBuffer;
-#endif
-
-
-
-float ComputeRealVertFOV() {
-	return 2.0f * atan2(0.5f * g_fCurInGameHeight, *g_fRawFOVDist) / DEG2RAD;
-}
-
-float ComputeRealHorzFOV() {
-	return 2.0f * atan2(0.5f * g_fCurInGameWidth, *g_fRawFOVDist) / DEG2RAD;
-}
-
-float RealVertFOVToRawFocalLength(float real_FOV_deg) {
-	return 0.5f * g_fCurInGameHeight / tan(0.5f * real_FOV_deg * DEG2RAD);
-}
-
-/**
- * Compute FOVscale and y_center for the hyperspace effect (and others that may need the FOVscale)
- */
-void ComputeHyperFOVParams() {
-	float y_center_raw, FOVscale_raw;
-	// Find the current center of the screen after it has been displaced by the cockpit camera.
-	// We only care about the y-coordinate, as we're going to use it along with the reticle to
-	// compute y_center.
-
-	// To find the y-center of the screen, we're going to use math. The tangent of the current
-	// pitch can give us the information right away. This is closely related to the FOV and
-	// is easy to see if you draw the focal length and the in-game height to form a right triangle.
-	float pitch = (float)PlayerDataTable[*g_playerIndex].cockpitCameraPitch / 65536.0f * 2.0f * PI;
-	float H = *g_fRawFOVDist * tan(pitch); // This will give us the height, in pixels, measured from the center of the screen
-	H += g_fCurInGameHeight / 2.0f;
-	log_debug("[DBG] [FOV] Screen Y-Center: %0.3f, ReticleCentroid: %0.3f, pitch: %0.3f, ", H, g_ReticleCentroid.y, pitch / DEG2RAD);
-	if (g_ReticleCentroid.y > -1.0f) {
-		// *sigh* for whatever stupid reason, sometimes we can have ReticleCentroid.y == 0.0 while looking straight ahead (pitch == 0)
-		// This situation completely destroys the calculation below, so we need to make sure that the camera pitch and the centroid
-		// are consistent.
-		// Another way to solve this problem is to prevent this calculation until the reticle centroid is close-ish to the center
-		// of the screen
-		float y_dist_from_screen_center = (float)fabs(g_ReticleCentroid.y - g_fCurInGameHeight / 2.0f);
-		if (y_dist_from_screen_center < g_fCurInGameHeight / 2.0f) {
-			// The reticle center visible this frame and it's not close to the edge of the screen.
-			// We can use it to compute y_center...
-
-			// The formula to compute y_center seems to be:
-			// (in-game-center - HUD_center) / in-game-height * 2.0f * comp_factor.
-			// The in-game-center has to be computer properly if the cockpit isn't facing forward
-			y_center_raw = 2.0f * (H - g_ReticleCentroid.y) / g_fCurInGameHeight;
-			log_debug("[DBG] [FOV] HUD_center to y_center: %0.3f", y_center_raw);
-			// We can stop looking for the reticle center now:
-			g_bYCenterHasBeenFixed = true;
-		}
-		else
-			log_debug("[DBG] [FOV] RETICLE COULD NOT BE USED COMPUTE Y_CENTER. WILL RETRY. Frame: %d", g_iPresentCounter);
-		// If the reticle center can't be used to compute y_center, then g_bYCenterHasBeenFixed will stay false, and we'll
-		// come back to this path on the next frame where a reticle is visible.
-	}
-	
-	if (!g_bYCenterHasBeenFixed) {
-		// Provide a default value if we couldn't compute y_center:
-		y_center_raw = 153.0f / g_fCurInGameHeight;
-		//y_center_raw = 0.0f; // I can't do this because the cockpits look wrong in the hangar. They look skewed
-	}
-	FOVscale_raw = 2.0f * *g_fRawFOVDist / g_fCurInGameHeight;
-
-	// Compute the aspect-ratio fix factors
-	float g_fWindowAspectRatio = max(1.0f, (float)g_WindowWidth / (float)g_WindowHeight);
-	if (g_bUseSteamVR)
-		g_fWindowAspectRatio = max(1.0f, (float)g_steamVRWidth / (float)g_steamVRHeight);
-	// The point where fixed and non-fixed params are about the same is given by the window aspect ratio
-	bool bFixFactors = g_fCurInGameAspectRatio > g_fWindowAspectRatio;
-	// The compensation factor is given by the ratio between the window aspect ratio and the in-game's 
-	// aspect ratio. In other words, the size of the display window will stretch the view, so we
-	// have to compensate for that.
-	float comp_factor = bFixFactors ? g_fWindowAspectRatio / g_fCurInGameAspectRatio : 1.0f;
-
-	log_debug("[DBG] [FOV] y_center raw: %0.3f, FOVscale raw: %0.3f, W/H: %0.0f, %0.0f, a/r: %0.3f, FIX: %d, comp_factor: %0.3f",
-		y_center_raw, FOVscale_raw, g_fCurInGameWidth, g_fCurInGameHeight, g_fCurInGameAspectRatio, bFixFactors, comp_factor);
-
-	// Compute the compensated FOV and y_center:
-	g_fFOVscale = comp_factor * FOVscale_raw;
-	g_fYCenter  = comp_factor * y_center_raw;
-
-	// Store the global FOVscale, y_center in the CB:
-	g_ShadertoyBuffer.FOVscale = g_fFOVscale;
-	g_ShadertoyBuffer.y_center = g_fYCenter;
-	
-	// If PreserveAspectRatio is 0, then we need to apply further compensation because the image
-	// will further stretch in either the X or Y axis.
-	g_ShadertoyBuffer.preserveAspectRatioComp[0] = 1.0f;
-	g_ShadertoyBuffer.preserveAspectRatioComp[1] = 1.0f;
-	// The VR modes behave just like PreserveAspectRatio = 0. The funny thing is that
-	// it will stretch the image in the same amount as when running in non-VR mode.
-	// This means that the aspect ratio of the in-game resolution should be an integer
-	// multiple of the Display/SteamVR window to avoid distortion.
-	if (!g_config.AspectRatioPreserved || g_bEnableVR)
-	{
-		float RealWindowAspectRatio = (float)g_WindowWidth / (float)g_WindowHeight;
-		// In SteamVR mode, the must compensate against the SteamVR window size
-		if (g_bUseSteamVR)
-			RealWindowAspectRatio = (float)g_steamVRWidth / (float)g_steamVRHeight;
-		
-		if (RealWindowAspectRatio > g_fCurInGameAspectRatio) {
-			// The display window is going to stretch the image horizontally, so we need
-			// to shrink the x axis:
-			if (RealWindowAspectRatio > 1.0f) // Make sure we shrink. If we divide by a value lower than 1, we'll stretch!
-				g_ShadertoyBuffer.preserveAspectRatioComp[0] = g_fCurInGameAspectRatio / RealWindowAspectRatio;
-			else
-				g_ShadertoyBuffer.preserveAspectRatioComp[0] = RealWindowAspectRatio / g_fCurInGameAspectRatio;
-		}
-		else {
-			// The display window is going to stretch the image vertically, so we need
-			// to shrink the y axis:
-			if (g_fCurInGameAspectRatio > 1.0f)
-				g_ShadertoyBuffer.preserveAspectRatioComp[1] = RealWindowAspectRatio / g_fCurInGameAspectRatio;
-			else
-				g_ShadertoyBuffer.preserveAspectRatioComp[1] = g_fCurInGameAspectRatio / RealWindowAspectRatio;
-		}
-
-		log_debug("[DBG] [FOV] Real Window a/r: %0.3f, preserveA/R-Comp: %0.3f, %0.3f",
-			RealWindowAspectRatio, g_ShadertoyBuffer.preserveAspectRatioComp[0], g_ShadertoyBuffer.preserveAspectRatioComp[1]);
-	}
-
-	// Compute the *real* vertical and horizontal FOVs:
-	g_fRealVertFOV = ComputeRealVertFOV();
-	g_fRealHorzFOV = ComputeRealHorzFOV();
-	// Compute the metric scale factor conversion
-	g_fOBJCurMetricScale = g_fCurInGameHeight * g_fOBJGlobalMetricMult / (SHADOW_OBJ_SCALE * 3200.0f);
-
-	// Populate the metric reconstruction CB. Noticeably, the Metric Rec values need the raw
-	// FOVscale and y_center, not the fixed ones. Making a 3D OBJ crosshairs as additional geometry
-	// also places the crosshairs at the right spot (the crosshairs need to be at (0,1,100))
-	g_MetricRecCBuffer.mr_FOVscale = FOVscale_raw;
-	g_MetricRecCBuffer.mr_y_center = y_center_raw;
-	g_MetricRecCBuffer.mr_cur_metric_scale = g_fOBJCurMetricScale;
-	g_MetricRecCBuffer.mr_aspect_ratio = g_fCurInGameAspectRatio;
-	g_MetricRecCBuffer.mr_z_metric_mult = g_fOBJ_Z_MetricMult;
-	g_MetricRecCBuffer.mr_shadow_OBJ_scale = SHADOW_OBJ_SCALE;
-	g_MetricRecCBuffer.mr_screen_aspect_ratio = g_fCurScreenWidth / g_fCurScreenHeight;
-	//g_MetricRecCBuffer.mr_vr_aspect_ratio_comp[0] = 1.0f;
-	//g_MetricRecCBuffer.mr_vr_aspect_ratio_comp[1] = 1.0f;
-	g_MetricRecCBuffer.mv_vr_vertexbuf_aspect_ratio_comp[0] = 1.0f;
-	g_MetricRecCBuffer.mv_vr_vertexbuf_aspect_ratio_comp[1] = 1.0f;
-	g_MetricRecCBuffer.mr_vr_aspect_ratio = 1.0f;
-	if (g_bEnableVR) {
-		if (g_bUseSteamVR) {
-			g_MetricRecCBuffer.mr_vr_aspect_ratio = (float)g_steamVRHeight / (float)g_steamVRWidth;
-		}
-		else {
-			// This is the DirectSBS mode. I don't have a reliable way to get the resolution of the 
-			// HMD device (which could be any cell phone + Google Cardboard for all we know). Instead,
-			// we need to trust that the user will set the current desktop resolution to the HMD's 
-			// resolution. So let's use the desktop's resolution to compensate for aspect ratio in
-			// SBS mode:
-			g_MetricRecCBuffer.mr_vr_aspect_ratio = (float)g_WindowHeight / (float)g_WindowWidth;
-		}
-	}
-
-	if (g_bEnableVR) {
-		if (g_bUseSteamVR) {
-			if (g_config.AspectRatioPreserved) {
-				g_MetricRecCBuffer.mv_vr_vertexbuf_aspect_ratio_comp[0] = 1.0f / g_ShadertoyBuffer.preserveAspectRatioComp[0];
-				g_MetricRecCBuffer.mv_vr_vertexbuf_aspect_ratio_comp[1] = 1.0f / g_ShadertoyBuffer.preserveAspectRatioComp[1];
-			}
-		}
-		else {
-			// DirectSBS path
-			g_MetricRecCBuffer.mv_vr_vertexbuf_aspect_ratio_comp[0] = 1.0f / g_ShadertoyBuffer.preserveAspectRatioComp[0];
-			g_MetricRecCBuffer.mv_vr_vertexbuf_aspect_ratio_comp[1] = 1.0f / g_ShadertoyBuffer.preserveAspectRatioComp[1];
-		}
-	}
-
-	// We just modified the Metric Reconstruction parameters, let's reapply them
-	g_bMetricParamsNeedReapply = true;
-
-	log_debug("[DBG] [FOV] Final y_center: %0.3f, FOV_Scale: %0.6f, RealVFOV: %0.2f, RealHFOV: %0.2f, mr_aspect_ratio: %0.3f, Frame: %d",
-		g_ShadertoyBuffer.y_center, g_ShadertoyBuffer.FOVscale, g_fRealVertFOV, g_fRealHorzFOV, 
-		g_MetricRecCBuffer.mr_aspect_ratio, g_iPresentCounter);
-
-	// DEBUG
-	//static bool bFirstTime = true;
-	//if (bFirstTime) {
-	//	g_fDebugFOVscale = g_ShadertoyBuffer.FOVscale;
-	//	//g_fDebugYCenter = g_ShadertoyBuffer.y_center;
-	//	bFirstTime = false;
-	//}
-	//g_fFOVscale = g_fDebugFOVscale;
-	//g_ShadertoyBuffer.FOVscale = g_fDebugFOVscale;
-	//
-	//g_fYCenter = g_fDebugYCenter;
-	//g_ShadertoyBuffer.y_center = g_fDebugYCenter;
-	////g_MetricRecCBuffer.mr_y_center = g_fDebugYCenter;
-	////log_debug("[DBG] [FOV] g_fDebugYCenter: %0.3f", g_fDebugYCenter);
-	//log_debug("[DBG] [FOV] g_fDebugYCenter: %0.3f, g_fDebugFOVscale: %0.6f", g_fDebugYCenter, g_fDebugFOVscale);
-	// DEBUG
-}
 
 // void capture()
 //#ifdef DBR_VR
@@ -7278,75 +6984,6 @@ void PrimarySurface::RenderSunFlare()
 	resources->InitInputLayout(resources->_inputLayout); // Not sure this is really needed
 }
 
-/*
- * Executes the action defined by "action" as per the Active Cockpit
- * definitions.
- */
-void ACRunAction(WORD *action) {
-	// Scan codes from: http://www.philipstorr.id.au/pcbook/book3/scancode.htm
-	// Scan codes: https://www.win.tue.nl/~aeb/linux/kbd/scancodes-1.html
-	// Based on code from: https://stackoverflow.com/questions/18647053/sendinput-not-equal-to-pressing-key-manually-on-keyboard-in-c
-	// Virtual key codes: https://docs.microsoft.com/en-us/windows/win32/inputdev/virtual-key-codes
-	// How to send extended scan codes
-	// https://stackoverflow.com/questions/36972524/winapi-extended-keyboard-scan-codes/36976260#36976260
-	// https://stackoverflow.com/questions/26283738/how-to-use-extended-scancodes-in-sendinput
-	INPUT input[MAX_AC_ACTION_LEN];
-	bool bEscapedAction = (action[0] == 0xe0);
-
-	if (action[0] == 0) { // void action, skip
-		//log_debug("[DBG] [AC] Skipping VOID action");
-		return;
-	}
-
-	// Special internal action: these actions don't need to synthesize any input
-	if (action[0] == 0xFF) {
-		switch (action[1]) {
-		case AC_HOLOGRAM_FAKE_VK_CODE:
-			g_bDCHologramsVisible = !g_bDCHologramsVisible;
-			return;
-		}
-		return;
-	}
-
-	//if (bEscapedAction)
-	//	log_debug("[DBG] [AC] Executing escaped code");
-	//log_debug("[DBG] [AC] Running action: ");
-	//DisplayACAction(action);
-
-	// Copy & initialize the scan codes
-	int i = 0, j = bEscapedAction ? 1 : 0;
-	while (action[j] && j < MAX_AC_ACTION_LEN) {
-		input[i].ki.wScan = action[j];
-		input[i].type = INPUT_KEYBOARD;
-		input[i].ki.time = 0;
-		input[i].ki.wVk = 0;
-		input[i].ki.dwExtraInfo = 0;
-		input[i].ki.dwFlags = KEYEVENTF_SCANCODE;
-		if (bEscapedAction) {
-			input[i].ki.dwFlags |= KEYEVENTF_EXTENDEDKEY;
-			//input[i].ki.dwExtraInfo = GetMessageExtraInfo();
-		}
-		i++; j++;
-	}
-
-	j = bEscapedAction ? 1 : 0;
-	while (action[j] && j < MAX_AC_ACTION_LEN) {
-		input[i].ki.wScan = action[j];
-		input[i].type = INPUT_KEYBOARD;
-		input[i].ki.time = 0;
-		input[i].ki.wVk = 0;
-		input[i].ki.dwExtraInfo = 0;
-		input[i].ki.dwFlags = KEYEVENTF_SCANCODE | KEYEVENTF_KEYUP;
-		if (bEscapedAction) {
-			input[i].ki.dwFlags |= KEYEVENTF_EXTENDEDKEY;
-			//input[i].ki.dwExtraInfo = GetMessageExtraInfo();
-		}
-		i++; j++;
-	}
-
-	// Send keydown/keyup events in one go: (this is the only way I found to enable the arrow/escaped keys)
-	SendInput(i, input, sizeof(INPUT));
-}
 
 /*
  * Input: offscreenBuffer (resolved here)
