+++ impl11/ddraw/SteamVR.cpp
@@ -20,6 +20,8 @@ const bool DEFAULT_STEAMVR_POS_FROM_FREEPIE = false;
 bool g_bInterleavedReprojection = DEFAULT_INTERLEAVED_REPROJECTION;
 bool g_bSteamVRDistortionEnabled = true;
 bool g_bSteamVRYawPitchRollFromMouseLook = false;
+bool g_bTogglePostPresentHandoff = false;
+bool g_bSteamVRMirrorWindowLeftEye = true;
 //bool g_bResetHeadCenter = true; // Reset the head center on startup
 //vr::HmdMatrix34_t g_EyeMatrixLeft, g_EyeMatrixRight;
 //Matrix4 g_EyeMatrixLeftInv, g_EyeMatrixRightInv;
