+++ impl11/ddraw/commonVR.h
@@ -26,8 +26,21 @@ extern bool g_bResetHeadCenter; // Reset the head center on startup
 extern float g_fVR_FOV;
 extern float g_fIPD;
 extern float g_fHalfIPD;
-// Set to true in PrimarySurface Present 2D (Flip)
-extern bool g_bInTechRoom;
+extern bool g_bInTechRoom; // Set to true in PrimarySurface Present 2D (Flip)
+
+extern float g_fPitchMultiplier, g_fYawMultiplier, g_fRollMultiplier;
+extern float g_fYawOffset, g_fPitchOffset;
+extern float g_fPosXMultiplier, g_fPosYMultiplier, g_fPosZMultiplier;
+extern float g_fMinPositionX, g_fMaxPositionX;
+extern float g_fMinPositionY, g_fMaxPositionY;
+extern float g_fMinPositionZ, g_fMaxPositionZ;
+extern float g_fFrameTimeRemaining;
+extern float g_fSteamVRMirrorWindow3DScale, g_fSteamVRMirrorWindowAspectRatio;
+extern Vector3 g_headCenter;
+extern bool g_bSteamVRPosFromFreePIE, g_bReshadeEnabled, g_bSteamVRDistortionEnabled, g_bSteamVRYawPitchRollFromMouseLook;
+extern vr::IVRSystem* g_pHMD;
+extern int g_iFreePIESlot, g_iSteamVR_Remaining_ms, g_iSteamVR_VSync_ms;
+extern bool g_bDirectSBSInitialized, g_bSteamVRInitialized;
 
 void EvaluateIPD(float NewIPD);
 void IncreaseIPD(float Delta);
