+++ impl11/ddraw/globals.h
@@ -9,20 +9,12 @@
 // METRIC RECONSTRUCTION:
 extern bool g_bYCenterHasBeenFixed;
 
-extern bool g_bCockpitPZHackEnabled;
-extern bool g_bOverrideAspectRatio;
-extern bool g_bEnableVR;
-extern TrackerType g_TrackerType;
-
 extern bool g_bDisableBarrelEffect, g_bEnableVR, g_bResetHeadCenter, g_bBloomEnabled, g_bAOEnabled, g_bCustomFOVApplied;
-extern bool g_bDirectSBSInitialized, g_bSteamVRInitialized, g_bClearHUDBuffers, g_bDCManualActivate, g_bGlobalDebugFlag;
+extern bool g_bClearHUDBuffers;
 extern bool g_bDumpSSAOBuffers, g_bEnableSSAOInShader, g_bEnableIndirectSSDO, g_bResetDC, g_bProceduralSuns, g_bEnableHeadLights;
 extern bool g_bShowSSAODebug, g_bShowNormBufDebug, g_bFNEnable, g_bShadowEnable, g_bGlobalSpecToggle, g_bToggleSkipDC;
 extern Vector4 g_LightVector[2];
 extern float g_fSpecIntensity, g_fSpecBloomIntensity, g_fFocalDist, g_fFakeRoll;
-
-extern bool g_bHDREnabled, g_bEdgeDetectorEnabled;
-extern float g_fHDRWhitePoint;
 
 extern bool g_bRendering3D; // Used to distinguish between 2D (Concourse/Menus) and 3D rendering (main in-flight game)
 // g_fZOverride is activated when it's greater than -0.9f, and it's used for bracket rendering so that 
@@ -74,20 +66,7 @@ extern bool g_bResetDC;
 extern int g_iReactorExplosionCount;
 
 
-// VR
-extern float g_fPitchMultiplier, g_fYawMultiplier, g_fRollMultiplier;
-extern float g_fYawOffset, g_fPitchOffset;
-extern float g_fPosXMultiplier, g_fPosYMultiplier, g_fPosZMultiplier;
-extern float g_fMinPositionX, g_fMaxPositionX;
-extern float g_fMinPositionY, g_fMaxPositionY;
-extern float g_fMinPositionZ, g_fMaxPositionZ;
-extern float g_fFrameTimeRemaining;
-extern float g_fSteamVRMirrorWindow3DScale, g_fSteamVRMirrorWindowAspectRatio;
-extern Vector3 g_headCenter;
-extern bool g_bResetHeadCenter, g_bSteamVRPosFromFreePIE, g_bReshadeEnabled, g_bSteamVRDistortionEnabled, g_bSteamVRYawPitchRollFromMouseLook;
-extern vr::IVRSystem* g_pHMD;
-extern int g_iFreePIESlot, g_iSteamVR_Remaining_ms, g_iSteamVR_VSync_ms;
-extern Matrix4 g_FullProjMatrixLeft, g_FullProjMatrixRight;
+
 
 
 extern HyperspacePhaseEnum g_HyperspacePhaseFSM;
@@ -107,9 +86,7 @@ extern Vector4 g_TempLightColor[2], g_TempLightVector[2];
 
 extern bool g_bFXAAEnabled;
 
-// BLOOM
-extern bool /* g_bDumpBloomBuffers, */ g_bDCManualActivate;
-extern BloomConfig g_BloomConfig;
+
 
 // SSAO
 extern SSAOTypeEnum g_SSAO_Type;
@@ -125,47 +102,6 @@ extern float g_fMoireOffsetDir, g_fMoireOffsetInd;
 
 extern Vector2 g_TriangleCentroid;
 extern float g_fTrianglePointerDist;
-
-// ACTIVE COCKPIT
-extern bool g_bActiveCockpitEnabled, g_bACActionTriggered, g_bACLastTriggerState, g_bACTriggerState;
-extern bool g_bFreePIEInitialized, g_bOriginFromHMD, g_bCompensateHMDRotation, g_bCompensateHMDPosition, g_bFreePIEControllerButtonDataAvailable;
-extern Vector4 g_contOriginWorldSpace, g_contOriginViewSpace, g_contDirWorldSpace, g_contDirViewSpace;
-extern Vector3 g_LaserPointer3DIntersection;
-extern float g_fBestIntersectionDistance, g_fLaserPointerLength;
-extern int g_iFreePIESlot, g_iFreePIEControllerSlot;
-extern float g_fContMultiplierX, g_fContMultiplierY, g_fContMultiplierZ, g_fFakeRoll;
-extern int g_iBestIntersTexIdx;
-extern ac_element g_ACElements[MAX_AC_TEXTURES_PER_COCKPIT];
-extern int g_iNumACElements, g_iLaserDirSelector;
-
-/*********************************************************/
-// DYNAMIC COCKPIT
-extern char g_sCurrentCockpit[128];
-extern DCHUDRegions g_DCHUDRegions;
-extern DCElemSrcBoxes g_DCElemSrcBoxes;
-extern dc_element g_DCElements[MAX_DC_SRC_ELEMENTS];
-
-extern float g_fCoverTextureBrightness;
-extern float g_fDCBrightness;
-extern int g_iNumDCElements;
-extern move_region_coords g_DCMoveRegions;
-extern float g_fCurInGameWidth, g_fCurInGameHeight, g_fCurInGameAspectRatio, g_fCurScreenWidth, g_fCurScreenHeight, g_fCurScreenWidthRcp, g_fCurScreenHeightRcp;
-extern bool g_bDCManualActivate, g_bDCApplyEraseRegionCommands, g_bReRenderMissilesNCounterMeasures;
-extern bool g_bGlobalDebugFlag, g_bInhibitCMDBracket;
-extern bool g_bHUDVisibleOnStartup;
-extern bool g_bCompensateFOVfor1920x1080;
-extern bool g_bDCWasClearedOnThisFrame;
-extern int g_iHUDOffscreenCommandsRendered;
-extern bool g_bEdgeEffectApplied;
-extern int g_WindowWidth, g_WindowHeight;
-extern float4 g_DCTargetingColor, g_DCWireframeLuminance;
-extern float4 g_DCTargetingIFFColors[6];
-extern float g_DCWireframeContrast;
-extern float g_fReticleScale;
-extern Vector2 g_SubCMDBracket; // Populated in XwaDrawBracketHook for the sub-CMD bracket when the enhanced 2D renderer is on
-// HOLOGRAMS
-extern float g_fDCHologramFadeIn, g_fDCHologramFadeInIncr, g_fDCHologramTime;
-extern bool g_bDCHologramsVisible, g_bDCHologramsVisiblePrev;
 
 // DEBUG vars
 extern Vector3 g_debug_v0, g_debug_v1, g_debug_v2;
@@ -198,20 +134,8 @@ extern int g_iSpeedShaderMaxParticles;
 //Vector4 g_prevFs(0, 0, 0, 0), g_prevUs(0, 0, 0, 0); //Not used
 extern D3DTLVERTEX g_SpeedParticles2D[MAX_SPEED_PARTICLES * 12];
 
-extern VertexShaderCBuffer g_VSCBuffer;
-extern PixelShaderCBuffer g_PSCBuffer;
-extern DCPixelShaderCBuffer g_DCPSCBuffer;
-extern ShadowMapVertexShaderMatrixCB g_ShadowMapVSCBuffer;
-extern MetricReconstructionCB g_MetricRecCBuffer;
-extern float g_fAspectRatio, g_fGlobalScale, g_fBrightness, g_fGUIElemsScale, g_fHUDDepth, g_fFloatingGUIDepth;
-extern float g_fCurScreenWidth, g_fCurScreenHeight, g_fCurInGameAspectRatio, g_fCurScreenWidthRcp, g_fCurScreenHeightRcp;
-extern float g_fCurInGameWidth, g_fCurInGameHeight, g_fMetricMult;
-extern int g_WindowWidth, g_WindowHeight;
-extern D3D11_VIEWPORT g_nonVRViewport;
 
-// MATERIALS
-extern bool g_bReloadMaterialsEnabled;
-extern Material g_DefaultGlobalMaterial;
+
 extern std::vector<ColorLightPair> g_TextureVector;
 /*
  * Used to store a list of textures for fast lookup. For instance, all suns must
@@ -239,7 +163,6 @@ extern float g_fDefaultFOVDist;
 extern float g_fDebugFOVscale, g_fDebugYCenter;
 extern float g_fCurrentShipFocalLength, g_fCurrentShipLargeFocalLength, g_fReticleScale;
 extern bool g_bYCenterHasBeenFixed;
-extern bool g_bTogglePostPresentHandoff;
 
 extern uint32_t* g_rawFOVDist; /* = (uint32_t *)0x91AB6C*/ // raw FOV dist(dword int), copy of one of the six values hard-coded with the resolution slots, which are what xwahacker edits
 extern float* g_fRawFOVDist; /*= (float *)0x8B94CC;*/ // FOV dist(float), same value as above
